easy_backend/requirements.txt
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.5